Added remote state to prompt
get_remote_state returns a string similar to the result of `git status -sb` and appends it to the existing cmder prompt. This allows you to see at a glance, not only which branch you are on and its dirty state, but also the number of commits ahead and behind the remote repository you are. Here's a screenshot of it in action - http://i.imgur.com/mk0ahik.png